Rename main file
